#188 Added IDepdendencyResolver.Resolve<TDependency>() method and removed DependencyResolverExtensions type
