+++ src/LightBDD.Core/Dependencies/IDependencyResolver.cs
@@ -13,5 +13,12 @@ namespace LightBDD.Core.Dependencies
         /// <param name="type">Dependency type.</param>
         /// <returns>Resolved dependency.</returns>
         object Resolve(Type type);
+
+        /// <summary>
+        /// Resolves dependency of the specified type.
+        /// </summary>
+        /// <typeparam name="TDependency">Type of dependency.</typeparam>
+        /// <returns>Resolved instance.</returns>
+        public TDependency Resolve<TDependency>();
     }
 }
