+++ src/LightBDD.Core/Dependencies/Implementation/DependencyContainer.cs
@@ -7,28 +7,22 @@ namespace LightBDD.Core.Dependencies.Implementation;
 
 internal class DependencyContainer : IDependencyContainer
 {
-    private readonly DependencyResolverProvider _resolverProvider;
+    public readonly DependencyResolverProvider ResolverProvider;
     private readonly IServiceProvider _serviceProvider;
     private readonly IDependencyResolver _resolver;
 
     public DependencyContainer(IServiceCollection collection)
     {
-        _resolverProvider = new(this);
+        ResolverProvider = new(this);
         collection.AddTransient<IDependencyResolver, DependencyResolver>();
-        collection.AddSingleton<IDependencyResolverProvider>(_resolverProvider);
+        collection.AddSingleton<IDependencyResolverProvider>(ResolverProvider);
         _serviceProvider = collection.BuildServiceProvider(true);
         _resolver = _serviceProvider.GetRequiredService<IDependencyResolver>();
     }
 
-    public object Resolve(Type type)
-    {
-        return _resolver.Resolve(type);
-    }
-
-    public IDependencyContainer BeginScope()
-    {
-        return new ScopedDependencyContainer(this, this, _serviceProvider.CreateAsyncScope());
-    }
+    public object Resolve(Type type) => _resolver.Resolve(type);
+    public TDependency Resolve<TDependency>() => _resolver.Resolve<TDependency>();
+    public IDependencyContainer BeginScope() => new ScopedDependencyContainer(this, this, _serviceProvider.CreateAsyncScope());
 
     public ValueTask DisposeAsync()
     {
@@ -37,35 +31,5 @@ internal class DependencyContainer : IDependencyContainer
 
         (_serviceProvider as IDisposable)?.Dispose();
         return default;
-    }
-
-    private class ScopedDependencyContainer : IDependencyContainer
-    {
-        private readonly AsyncServiceScope _scope;
-        private readonly DependencyContainer _root;
-        private readonly IDependencyContainer _parent;
-        private readonly IDependencyResolver _resolver;
-
-        public ScopedDependencyContainer(DependencyContainer root, IDependencyContainer parent, AsyncServiceScope scope)
-        {
-            _root = root;
-            _parent = parent;
-            _scope = scope;
-            _root._resolverProvider.SetCurrent(this);
-            _resolver = scope.ServiceProvider.GetRequiredService<IDependencyResolver>();
-        }
-
-        public async ValueTask DisposeAsync()
-        {
-            _root._resolverProvider.SetCurrent(_parent);
-            await _scope.DisposeAsync();
-        }
-
-        public object Resolve(Type type) => _resolver.Resolve(type);
-
-        public IDependencyContainer BeginScope()
-        {
-            return new ScopedDependencyContainer(_root, this, _scope.ServiceProvider.CreateAsyncScope());
-        }
     }
 }
